update
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@5218 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ ext/libsamplerate/src_linear.c
@@ -79,7 +79,7 @@ linear_vari_process (SRC_PRIVATE *psrc, SRC_DATA *data)
 
 		for (ch = 0 ; ch < priv->channels ; ch++)
 		{	data->data_out [priv->out_gen] = (float) (priv->last_value [ch] + input_index *
-										(data->data_in [ch] - priv->last_value [ch])) ;
+										((double) data->data_in [ch] - priv->last_value [ch])) ;
 			priv->out_gen ++ ;
 			} ;
 
@@ -104,7 +104,7 @@ linear_vari_process (SRC_PRIVATE *psrc, SRC_DATA *data)
 
 		for (ch = 0 ; ch < priv->channels ; ch++)
 		{	data->data_out [priv->out_gen] = (float) (data->data_in [priv->in_used - priv->channels + ch] + input_index *
-						(data->data_in [priv->in_used + ch] - data->data_in [priv->in_used - priv->channels + ch])) ;
+						((double) data->data_in [priv->in_used + ch] - data->data_in [priv->in_used - priv->channels + ch])) ;
 			priv->out_gen ++ ;
 			} ;
